src/back-end/web/requirements.txt
amqp==5.1.1
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
astroid==2.11.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.1.0
boto3==1.21.24
botocore==1.24.24
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
cfgv==3.2.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.1
daphne==3.0.2
dill==0.3.4
distlib==0.3.1
Django==4.0.4
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-guardian==2.4.0
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-spectacular==0.22.0
drf-yasg==1.20.0
execnet==1.9.0
filelock==3.0.12
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
hyperlink==21.0.0
identify==2.4.12
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.7.0
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
jsonschema==4.4.0
kombu==5.2.4
lazy-object-proxy==1.5.2
Markdown==3.3.6
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.5.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.11.1
prompt-toolkit==3.0.29
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pylint==2.13.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymongo==3.12.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.0
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.2.4
toml==0.10.2
tomli==2.0.1
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
virtualenv==20.4.3
watchgod==0.8
wcwidth==0.2.5
websockets==10.2
wrapt==1.12.1
zipp==3.7.0
zope.interface==5.4.0
